requirements.txt
psutil>=5.9.0  # For memory monitoring
scipy>=1.11.0
tqdm>=4.65.0
streamlit>=1.28.0
pandas>=2.0.0
altair>=5.0.0
numpy>=1.24.0
PyYAML>=6.0
python-docx>=0.8.11
pdfkit>=1.0.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
streamlit-autorefresh>=0.0.1
pytest>=7.0.0
